# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Coaching planner backend Documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "rubenvdkamp08@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Coaching planner API Server"
        },
        {
            "url": "http://my-default-host.com",
            "description": "Coaching planner API Server"
        },
        {
            "url": "http://my-default-host.com",
            "description": "Coaching planner API Server"
        },
        {
            "url": "http://coachingplanner.test"
        }
    ],
    "paths": {
...
        "/students": {
            "get": {
                "tags": [
                    "Student"
                ],
                "summary": "Get list of students",
                "description": "Returns list of coaches students",
                "operationId": "get_students",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    [
                        "coachingplanner_auth_key"
                    ]
                ]
            }
        },
        "/student": {
            "post": {
                "tags": [
                    "Student"
                ],
                "summary": "Create new student",
                "description": "Created a new student for the coach that can assign coaching moment",
                "operationId": "create_student",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateStudentRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Student is successfully created"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Something went wrong"
                    }
                },
                "security": [
                    [
                        "coachingplanner_auth_key"
                    ]
                ]
            }
        },
        "/student/{student}": {
            "get": {
                "tags": [
                    "Student"
                ],
                "summary": "Get details of specific student",
                "description": "Returns details of a specific student",
                "operationId": "get_student",
                "parameters": [
                    {
                        "name": "student",
                        "in": "query",
                        "description": "guid of the student",
                        "required": true,
                        "style": "form",
                        "schema": {
                            "type": "string",
                            "example": "9c058e90-7ef1-11ea-a2bd-6db2999910ca"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    [
                        "coachingplanner_auth_key"
                    ]
                ]
            },
            "put": {
                "tags": [
                    "Student"
                ],
                "summary": "Edit the given student when it's a student of the coach",
                "description": "Edit the given student",
                "operationId": "edit_student",
                "parameters": [
                    {
                        "name": "student",
                        "in": "query",
                        "description": "guid of the student",
                        "required": true,
                        "style": "form",
                        "schema": {
                            "type": "string",
                            "example": "9c058e90-7ef1-11ea-a2bd-6db2999910ca"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateStudentRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "500": {
                        "description": "Something went wrong, please try again."
                    }
                },
                "security": [
                    [
                        "coachingplanner_auth_key"
                    ]
                ]
            },
            "delete": {
                "tags": [
                    "Student"
                ],
                "summary": "Delete the given student when it's a student of the coach",
                "description": "Delete the given student",
                "operationId": "delete_student",
                "parameters": [
                    {
                        "name": "student",
                        "in": "query",
                        "description": "guid of the student",
                        "required": true,
                        "style": "form",
                        "schema": {
                            "type": "string",
                            "example": "9c058e90-7ef1-11ea-a2bd-6db2999910ca"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "500": {
                        "description": "Something went wrong, please try again."
                    }
                },
                "security": [
                    [
                        "coachingplanner_auth_key"
                    ]
                ]
            }
        },
        "/student/import": {
            "post": {
                "tags": [
                    "Student"
                ],
                "summary": "Import students from .csv",
                "description": "Import student from csv to make it easier to add students",
                "operationId": "import_students",
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Something went wrong"
                    }
                },
                "security": [
                    [
                        "coachingplanner_auth_key"
                    ]
                ]
            }
        },
        "/user/students/{user}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get list of students",
                "description": "Returns list of students of the coach",
                "operationId": "get_students_of_coach",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    [
                        "coachingplanner_auth_key"
                    ]
                ]
            }
        },
        "/user/{user}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get Coach (user) details",
                "description": "Returns details of the user (coach)",
                "operationId": "get_user_details",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    [
                        "coachingplanner_auth_key"
                    ]
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "CreateStudentRequest": {
                "title": "Create student request",
                "description": "Create new student request body data",
                "required": [
                    "studentnumber",
                    "firstname",
                    "lastname",
                    "email"
                ],
                "properties": {
                    "studentnumber": {
                        "title": "studentnumber",
                        "description": "Studentnumber of the new student",
                        "type": "string",
                        "example": "1112200"
                    },
                    "firstname": {
                        "title": "firstname",
                        "description": "Firstname of the new student",
                        "type": "string",
                        "example": "Peter"
                    },
                    "suffix": {
                        "title": "suffix",
                        "description": "Suffix of the new student",
                        "type": "string",
                        "example": "de"
                    },
                    "lastname": {
                        "title": "lastname",
                        "description": "Lastname of the new student",
                        "type": "string",
                        "example": "Johnson"
                    },
                    "email": {
                        "title": "email",
                        "description": "School mailof the new student",
                        "type": "string",
                        "example": "s1112200@student.windesheim.nl"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "coachingplanner_auth_key": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}